// nextjs/components/database/schemas/public/MeasureValue.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.measure_value */
export type Measure_valueName = string;

/** Identifier type for public.measure_value */
export type Measure_valueCompany = string;

/** Represents the table public.measure_value */
export default interface MeasureValue {
  name: Measure_valueName;

  company: Measure_valueCompany;

  translations: unknown;

  measureUnit: string | null;

  type: string;

  isEditable: boolean;

  defaultValue: string | null;
}

/** Represents the initializer for the table public.measure_value */
export interface MeasureValueInitializer {
  name: Measure_valueName;

  company: Measure_valueCompany;

  translations: unknown;

  measureUnit?: string | null;

  type: string;

  isEditable: boolean;

  defaultValue?: string | null;
}

/** Represents the mutator for the table public.measure_value */
export interface MeasureValueMutator {
  name?: Measure_valueName;

  company?: Measure_valueCompany;

  translations?: unknown;

  measureUnit?: string | null;

  type?: string;

  isEditable?: boolean;

  defaultValue?: string | null;
}